fix: added malloc fail checks and seperated few functions for norm proof
